// src/redux/types.ts
export enum authTypes {
  AUTHENTICATE_USER_SUCCESS = 'AUTHENTICATE_USER_SUCCESS',
  AUTHENTICATE_USER_ERROR = 'AUTHENTICATE_USER_ERROR',
  REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS',
  REGISTER_USER_ERROR = 'REGISTER_USER_ERROR',
  LOGOUT = 'LOGOUT',
}

export enum postTypes {
  LOADING_POSTS = 'LOADING_POSTS',
  CREATE_POST_SUCCESS = 'CREATE_POST_SUCCESS',
  CREATE_POST_ERROR = 'CREATE_POST_ERROR',
  GET_FEED_POST_SUCCESS = 'GET_FEED_POST_SUCCESS',
  GET_FEED_POST_ERROR = 'GET_FEED_POST_ERROR',
  LIKE_FEED_POST_SUCCESS = 'LIKE_FEED_POST_SUCCESS',
  LIKE_FEED_POST_ERROR = 'LIKE_FEED_POST_ERROR',
  DISLIKE_FEED_POST_SUCCESS = 'DISLIKE_FEED_POST_SUCCESS',
  DISLIKE_FEED_POST_ERROR = 'DISLIKE_FEED_POST_ERROR',
  GET_MY_POSTS_SUCCESS = 'GET_MY_POSTS_SUCCESS',
  GET_MY_POSTS_ERROR = 'GET_MY_POSTS_ERROR',
  DELETE_FEED_POST_SUCCESS = 'DELETE_FEED_POST_SUCCESS',
  DELETE_FEED_POST_ERROR = 'DELETE_FEED_POST_ERROR',
}

export enum userType {
  LOADING_USER = 'LOADING_USER',
  GET_USERS_SUCCESS = 'GET_USERS_SUCCESS',
  GET_USERS_ERROR = 'GET_USERS_ERROR',
  GET_PROFILE_SUCCESS = 'GET_PROFILE_SUCCESS',
  GET_PROFILE_ERROR = 'GET_PROFILE_ERROR',
}